consolidating everything to essentials.py again
