+++ src/interpreter/bl_types/exits.py
@@ -2,7 +2,7 @@
 
 from dataclasses import dataclass
 
-from .essentials import Exit, Value
+from .essentials import Exit
 
 
 # pylint: disable=too-few-public-methods
@@ -16,9 +16,3 @@ class Break(Exit):
 @dataclass(frozen=True)
 class Continue(Exit):
     """Continue statement"""
-
-
-@dataclass(frozen=True)
-class Return(Exit):
-    """Return statement"""
-    value: Value
