--- src/interpreter/env.py
@@ -1,77 +1,0 @@
-"""Interpreter environment"""
-
-
-from dataclasses import dataclass
-from typing import TYPE_CHECKING, cast
-
-from lark.tree import Meta
-
-from .bl_types import (
-    ExpressionResult, BLError, Value, Instance, VarNotFoundException
-)
-
-if TYPE_CHECKING:
-    from ..main import ASTInterpreter
-
-
-@dataclass
-class Var:
-    """Interpreter mutable binding"""
-
-    value: Value
-
-
-class Env:
-    """Interpreter environment"""
-
-    interpreter: "ASTInterpreter"
-    vars: dict[str, Var]
-    parent: "Env | None"
-
-    def __init__(
-        self, interpreter: "ASTInterpreter",
-        vars_: dict[str, Var] | None = None, parent: "Env | None" = None,
-    ):
-        if vars_ is None:
-            self.vars = {}
-        else:
-            self.vars = vars_
-        self.interpreter = interpreter
-        self.parent = parent
-
-    def new_var(self, name: str, value: Value) -> None:
-        """Set a new/existing variable"""
-        self.vars[name] = Var(value)
-
-    def get_var(self, name: str, meta: Meta | None) -> ExpressionResult:
-        """Retrieve the value of a variable"""
-        resolve_result = self.resolve_var(name, meta)
-        match resolve_result:
-            case Var(value=value):
-                return value
-            case BLError():
-                return resolve_result
-
-    def set_var(self, name: str, value: Value, meta: Meta | None
-                ) -> BLError | None:
-        """Assign to an existing variable name"""
-        resolve_result = self.resolve_var(name, meta)
-        match resolve_result:
-            case Var():
-                resolve_result.value = value
-            case BLError():
-                return resolve_result
-
-    def resolve_var(self, name: str, meta: Meta | None) -> Var | BLError:
-        """Resolve a variable name"""
-        if name in self.vars:
-            return self.vars[name]
-        if self.parent is not None:
-            return self.parent.resolve_var(name, meta)
-        return BLError(cast(
-            Instance, VarNotFoundException.new([], self.interpreter, meta)
-        ))
-
-    def copy(self) -> 'Env':
-        """Copy the environment (for capturing variables in closures)"""
-        return Env(self.interpreter, self.vars.copy(), self.parent)
